Simplify inheritance of `SubsequentOrderingAssertions`
